core: Fix storage size overflows on 32-bit targets
Storage sizes and progress values could wrap or truncate on 32-bit hosts, and device-controlled packets could over-allocate, so parse as u64, fail loudly on overflow, and cap packets.

+++ tui/src/cli/commands/device/download.rs
@@ -56,7 +56,7 @@ impl DeviceCommand for DownloadArgs {
             return Err(anyhow::anyhow!("Partition '{}' not found on device.", self.partition));
         };
 
-        if file_size > part.size as u64 {
+        if file_size > part.size {
             return Err(anyhow::anyhow!(
                 "File size ({}) exceeds partition size ({}).",
                 file_size,
